populate the inference tests file
no changes: code moved from core test file
