+++ test/core.jl
@@ -1022,33 +1022,6 @@ let
     @test g1632(:a, 2) == 1
 end
 
-# issue #1628
-type I1628{X}
-    x::X
-end
-let
-    # here the potential problem is that the run-time value of static
-    # parameter X in the I1628 constructor is (DataType,DataType),
-    # but type inference will track it more accurately as
-    # (Type{Integer}, Type{Int}).
-    f1628() = I1628((Integer,Int))
-    @test isa(f1628(), I1628{Tuple{DataType,DataType}})
-end
-
-let
-    fT{T}(x::T) = T
-    @test fT(Any) === DataType
-    @test fT(Int) === DataType
-    @test fT(Type{Any}) === DataType
-    @test fT(Type{Int}) === DataType
-
-    ff{T}(x::Type{T}) = T
-    @test ff(Type{Any}) === Type{Any}
-    @test ff(Type{Int}) === Type{Int}
-    @test ff(Any) === Any
-    @test ff(Int) === Int
-end
-
 # issue #2098
 let
     i2098() = begin
@@ -1157,18 +1130,6 @@ end
 
 @M2982.bad(T2982)
 @test T2982.super === M2982.U
-
-# issue #3182
-f3182{T}(::Type{T}) = 0
-f3182(x) = 1
-function g3182(t::DataType)
-    # tricky thing here is that DataType is a concrete type, and a
-    # subtype of Type, but we cannot infer the T in Type{T} just
-    # by knowing (at compile time) that the argument is a DataType.
-    # however the ::Type{T} method should still match at run time.
-    f3182(t)
-end
-@test g3182(Complex) == 0
 
 # issue #3221
 let x = fill(nothing, 1)
@@ -1679,32 +1640,6 @@ let
     @test Test()() === nothing
 end
 
-# issue #5906
-
-abstract Outer5906{T}
-
-immutable Inner5906{T}
-   a:: T
-end
-
-immutable Empty5906{T} <: Outer5906{T}
-end
-
-immutable Hanoi5906{T} <: Outer5906{T}
-    a::T
-    succ :: Outer5906{Inner5906{T}}
-    Hanoi5906(a) = new(a, Empty5906{Inner5906{T}}())
-end
-
-function f5906{T}(h::Hanoi5906{T})
-    if isa(h.succ, Empty5906) return end
-    f5906(h.succ)
-end
-
-# can cause infinite recursion in type inference via instantiation of
-# the type of the `succ` field
-@test f5906(Hanoi5906{Int}(1)) === nothing
-
 # make sure front end can correctly print values to error messages
 let
     ex = expand(parse("\"a\"=1"))
@@ -1835,13 +1770,6 @@ f6502() = convert(Tuple{Vararg{Int}}, (10,))
 @test f6502() === (10,)
 @test convert(Tuple{Bool,Vararg{Int}}, (true,10)) === (true,10)
 @test convert(Tuple{Int,Vararg{Bool}}, (true,1,0)) === (1,true,false)
-
-# issue on the flight from DFW
-# (type inference deducing Type{:x} rather than Symbol)
-type FooBarDFW{s}; end
-fooDFW(p::Type{FooBarDFW}) = string(p.parameters[1])
-fooDFW(p) = string(p.parameters[1])
-@test fooDFW(FooBarDFW{:x}) == "x" # not ":x"
 
 # issue #6611
 function crc6611(spec)
@@ -2028,13 +1956,6 @@ i7652()
 
 # issue #7679
 @test map(f->f(), Any[ ()->i for i=1:3 ]) == Any[1,2,3]
-
-# issue #7810
-type Foo7810{T<:AbstractVector}
-    v::T
-end
-bar7810() = [Foo7810([(a,b) for a in 1:2]) for b in 3:4]
-@test Base.return_types(bar7810,Tuple{})[1] == Array{Foo7810{Array{Tuple{Int,Int},1}},1}
 
 # issue 7897
 function issue7897!(data, arr)
@@ -2301,13 +2222,6 @@ function f9947()
 end
 @test f9947() == UInt128(1)
 
-# Type inference for tuple parameters
-immutable fooTuple{s}; end
-barTuple1() = fooTuple{(:y,)}()
-barTuple2() = fooTuple{tuple(:y)}()
-
-@test Base.return_types(barTuple1,Tuple{})[1] == Base.return_types(barTuple2,Tuple{})[1] == fooTuple{(:y,)}
-
 #issue #9835
 module M9835
     using Base.Test
@@ -2393,12 +2307,6 @@ f7221{T<:Number}(::T) = 1
 f7221(::BitArray) = 2
 f7221(::AbstractVecOrMat) = 3
 @test f7221(trues(1)) == 2
-
-# issue #9232
-arithtype9232{T<:Real}(::Type{T},::Type{T}) = arithtype9232(T)
-result_type9232{T1<:Number,T2<:Number}(::Type{T1}, ::Type{T2}) = arithtype9232(T1, T2)
-# this gave a "type too large", but not reliably
-@test length(code_typed(result_type9232, Tuple{Type{TypeVar(:_, Union{Float32,Float64})}, Type{TypeVar(:T2, Number)}})) == 1
 
 # test functionality of non-power-of-2 bitstype constants
 bitstype 24 Int24
@@ -2935,11 +2843,6 @@ gc()
 @test collect(enumerate((Tuple,Int))) == [(1,Tuple), (2,Int)]
 @test collect(enumerate((Tuple,3))) == [(1,Tuple), (2,3)]
 
-# issue #10878
-function g10878(x; kw...); end
-invoke_g10878() = invoke(g10878, Tuple{Any}, 1)
-@code_typed invoke_g10878()
-
 # issue #10978
 typealias TupleType10978{T<:Tuple} Type{T}
 f10978(T::TupleType10978) = isa(T, TupleType10978)
@@ -3004,10 +2907,6 @@ let a = [Pair(1,2), Pair("a","b")]
     @test typeof(a) == Vector{Pair}
     @test typeof(a) <: Vector{Pair}
 end
-
-# issue #11366
-f11366{T}(x::Type{Ref{T}}) = Ref{x}
-@test !isleaftype(Base.return_types(f11366, (Any,))[1])
 
 # issue #11065, #1571
 function f11065()
@@ -3164,17 +3063,6 @@ end
 @test_throws ErrorException NTuple{-1, Int}
 @test_throws TypeError Union{Int, 1}
 
-# issue #10930
-@test isa(code_typed(promote,(Any,Any,Vararg{Any})), Array)
-find_tvar10930{T<:Tuple}(sig::Type{T}) = 1
-function find_tvar10930(arg)
-    if arg<:Tuple
-        find_tvar10930(arg[random_var_name])
-    end
-    return 1
-end
-@test find_tvar10930(Vararg{Int}) === 1
-
 # issue #12003
 const DATE12003 = DateTime(1917,1,1)
 failure12003(dt=DATE12003) = Dates.year(dt)
@@ -3287,13 +3175,6 @@ let x = (1,2)
     @test f12517() === Val{(1,2)}
 end
 
-# issue #12476
-function f12476(a)
-    (k, v) = a
-    v
-end
-@inferred f12476(1.0 => 1)
-
 # don't allow Vararg{} in Union{} type constructor
 @test_throws TypeError Union{Int,Vararg{Int}}
 
@@ -3307,29 +3188,6 @@ end
 typealias PossiblyInvalidUnion{T} Union{T,Int}
 @test_throws TypeError PossiblyInvalidUnion{1}
 
-# issue #12551 (make sure these don't throw in inference)
-Base.return_types(unsafe_load, (Ptr{nothing},))
-Base.return_types(getindex, (Vector{nothing},))
-
-# issue #12636
-module MyColors
-
-abstract Paint{T}
-immutable RGB{T<:AbstractFloat} <: Paint{T}
-    r::T
-    g::T
-    b::T
-end
-
-myeltype{T}(::Type{Paint{T}}) = T
-myeltype{P<:Paint}(::Type{P}) = myeltype(supertype(P))
-myeltype(::Type{Any}) = Any
-
-end
-
-@test @inferred(MyColors.myeltype(MyColors.RGB{Float32})) == Float32
-@test @inferred(MyColors.myeltype(MyColors.RGB)) == Any
-
 # issue #12569
 @test_throws ArgumentError symbol("x"^10_000_000)
 @test_throws ArgumentError gensym("x"^10_000_000)
@@ -3338,10 +3196,6 @@ end
 
 # meta nodes for optional positional arguments
 @test Base.uncompressed_ast(expand(:(@inline f(p::Int=2) = 3)).args[2].args[3]).args[3].args[1].args[1] === :inline
-
-# issue #12826
-f12826{I<:Integer}(v::Vector{I}) = v[1]
-@test Base.return_types(f12826,Tuple{Array{TypeVar(:I, Integer),1}})[1] == Integer
 
 # issue #13007
 call13007{T,N}(::Type{Array{T,N}}) = 0
@@ -3579,16 +3433,6 @@ function __f_isa_arg_1()
 end
 @test __f_isa_arg_1() == 1
 
-# non-terminating inference, issue #14009
-type A14009{T}; end
-A14009{T}(a::T) = A14009{T}()
-f14009(a) = rand(Bool) ? f14009(A14009(a)) : a
-code_typed(f14009, (Int,))
-
-type B14009{T}; end
-g14009(a) = g14009(B14009{a})
-code_typed(g14009, (Type{Int},))
-
 # issue #14477
 immutable Z14477
     fld::Z14477
@@ -3742,10 +3586,6 @@ let grphtest = ((1, [2]),)
     end
 end
 
-let f(T) = Type{T}
-    @test Base.return_types(f, Tuple{Type{Int}}) == [Type{Type{Int}}]
-end
-
 # issue #13229
 module I13229
     using Base.Test
@@ -3757,18 +3597,6 @@ module I13229
         global z = f(i)
     end
     @test z == 10
-end
-
-# issue #12474
-@generated function f12474(::Any)
-    :(for i in 1
-      end)
-end
-let
-    ast12474 = code_typed(f12474, Tuple{Float64})
-    for (_, vartype) in ast12474[1].args[2][1]
-        @test isleaftype(vartype)
-    end
 end
 
 # issue #15186
@@ -3845,23 +3673,6 @@ let ary = Vector{Any}(10)
         check_undef_and_fill(ary, 1:(2n + 4))
     end
 end
-
-# pr #15259
-immutable A15259
-    x
-    y
-end
-# check that allocation was ellided
-@eval f15259(x,y) = (a = $(Expr(:new, :A15259, :x, :y)); (a.x, a.y, getfield(a,1), getfield(a, 2)))
-@test isempty(filter(x -> isa(x,Expr) && x.head === :(=) &&
-                          isa(x.args[2], Expr) && x.args[2].head === :new,
-                     code_typed(f15259, (Any,Int))[1].args[3].args))
-@test f15259(1,2) == (1,2,1,2)
-# check that error cases are still correct
-@eval g15259(x,y) = (a = $(Expr(:new, :A15259, :x, :y)); a.z)
-@test_throws ErrorException g15259(1,1)
-@eval h15259(x,y) = (a = $(Expr(:new, :A15259, :x, :y)); getfield(a, 3))
-@test_throws BoundsError h15259(1,1)
 
 # issue #15283
 j15283 = 0
